Conflicts:
	src/main/java/net/pms/configuration/PmsConfiguration.java
	src/main/java/net/pms/configuration/RendererConfiguration.java
	src/main/java/net/pms/dlna/DLNAMediaInfo.java
	src/main/java/net/pms/encoders/FFMpegVideo.java
	src/main/java/net/pms/formats/v2/SubtitleUtils.java

+++ src/main/java/net/pms/dlna/DVDISOTitle.java
@@ -280,7 +280,7 @@ public class DVDISOTitle extends DLNAResource {
 	public long length(RendererConfiguration mediaRenderer) {
 		// WDTV Live at least, needs a realistic size for stop/resume to works proberly. 2030879 = ((15000 + 256) * 1024 / 8 * 1.04) : 1.04 = overhead
 		int cbr_video_bitrate = getDefaultRenderer().getCBRVideoBitrate();
-		return (cbr_video_bitrate > 0) ? (long) (((cbr_video_bitrate + 256) * 1024 / 8 * 1.04) * getMedia().getDurationInSeconds()) : length();
+		return (cbr_video_bitrate > 0) ? (long) (((cbr_video_bitrate + 256) * 1024 / (double) 8 * 1.04) * getMedia().getDurationInSeconds()) : length();
 	}
 
 	@Override
